Moving API-related tests into their own directory and splitting into multiple modules
