+++ tardis/tardis_portal/tests/api/test_datafile_metadata_resources.py
@@ -1,0 +1,16 @@
+'''
+Testing the DatafileParameter and DatafileParameterSet resources in
+MyTardis's Tastypie-based REST API
+
+.. moduleauthor:: Grischa Meyer <grischa@gmail.com>
+.. moduleauthor:: James Wettenhall <james.wettenhall@monash.edu>
+'''
+from . import MyTardisResourceTestCase
+
+
+class DatafileParameterSetResourceTest(MyTardisResourceTestCase):
+    pass
+
+
+class DatafileParameterResourceTest(MyTardisResourceTestCase):
+    pass
