+++ tardis/tardis_portal/tests/api/test_dataset_metadata_resources.py
@@ -1,0 +1,16 @@
+'''
+Testing the DatasetParameter and DatasetParameterSet resources in
+MyTardis's Tastypie-based REST API
+
+.. moduleauthor:: Grischa Meyer <grischa@gmail.com>
+.. moduleauthor:: James Wettenhall <james.wettenhall@monash.edu>
+'''
+from . import MyTardisResourceTestCase
+
+
+class DatasetParameterSetResourceTest(MyTardisResourceTestCase):
+    pass
+
+
+class DatasetParameterResourceTest(MyTardisResourceTestCase):
+    pass
