+++ tardis/tardis_portal/tests/api/test_experiment_metadata_resources.py
@@ -1,0 +1,16 @@
+'''
+Testing the ExperimentParameter and ExperimentParameterSet resources in
+MyTardis's Tastypie-based REST API
+
+.. moduleauthor:: Grischa Meyer <grischa@gmail.com>
+.. moduleauthor:: James Wettenhall <james.wettenhall@monash.edu>
+'''
+from . import MyTardisResourceTestCase
+
+
+class ExperimentParameterSetResourceTest(MyTardisResourceTestCase):
+    pass
+
+
+class ExperimentParameterResourceTest(MyTardisResourceTestCase):
+    pass
