+++ tardis/tardis_portal/tests/api/test_location_resource.py
@@ -1,0 +1,14 @@
+'''
+Testing the Location resource in MyTardis's Tastypie-based REST API
+
+"Location" was the old name for what is now known as a StorageBox
+in MyTardis.  The API v1 endpoint is still /api/v1/location/
+
+.. moduleauthor:: Grischa Meyer <grischa@gmail.com>
+.. moduleauthor:: James Wettenhall <james.wettenhall@monash.edu>
+'''
+from . import MyTardisResourceTestCase
+
+
+class LocationResourceTest(MyTardisResourceTestCase):
+    pass
