+++ tardis/tardis_portal/tests/api/test_replica_resource.py
@@ -1,0 +1,14 @@
+'''
+Testing the Replica resource in MyTardis's Tastypie-based REST API
+
+"Replica" was the old name for what is now known as a DataFileObject
+in MyTardis.  The API v1 endpoint is still /api/v1/replica/
+
+.. moduleauthor:: Grischa Meyer <grischa@gmail.com>
+.. moduleauthor:: James Wettenhall <james.wettenhall@monash.edu>
+'''
+from . import MyTardisResourceTestCase
+
+
+class ReplicaResourceTest(MyTardisResourceTestCase):
+    pass
