+++ tardis/tardis_portal/tests/api/test_schema_resource.py
@@ -1,0 +1,12 @@
+'''
+Testing the Schema resource in MyTardis's Tastypie-based REST API
+
+.. moduleauthor:: Grischa Meyer <grischa@gmail.com>
+.. moduleauthor:: James Wettenhall <james.wettenhall@monash.edu>
+
+'''
+from . import MyTardisResourceTestCase
+
+
+class SchemaResourceTest(MyTardisResourceTestCase):
+    pass
